+++ openapi-config.yaml
@@ -11,10 +11,10 @@ servers:
 #  url: http://127.0.0.1:5000
 termsOfService: http://robokop.renci.org:7055/tos?service_long=ARAGORN&provider_long=RENCI
 title: ARAGORN
+version:
 tags:
 - name: translator
 - name: ARA
-- name: ARAGORN
 - name: trapi
 x-translator:
   biolink-version: 2.1.0
